refactor:schedule-data
스케줄 데이터에 프로필 이미지를 추가했습니다.
