+++ Back-End/src/main/java/net/scit/backend/member/service/MemberService.java
@@ -42,6 +42,4 @@ public interface MemberService {
 
     //로그인 상태 조회
     MemberLoginStatusDTO getLoginStatus(String userEmail);
-
-    ResultDTO<SuccessDTO>linkAccount(String email, boolean linkYn);
 }
