+++ Back-End/src/main/java/net/scit/backend/oauth/controller/OAuthController.java
@@ -2,14 +2,12 @@ package net.scit.backend.oauth.controller;
 
 import lombok.RequiredArgsConstructor;
 import net.scit.backend.common.ResultDTO;
+import net.scit.backend.common.SuccessDTO;
 import net.scit.backend.member.dto.TokenDTO;
 import net.scit.backend.oauth.dto.GoogleDTO;
 import net.scit.backend.oauth.service.OAuthService;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequiredArgsConstructor
@@ -21,6 +19,17 @@ public class OAuthController {
     @PostMapping("/google")
     public ResponseEntity<ResultDTO<TokenDTO>> google(@RequestBody GoogleDTO googleDTO) {
         ResultDTO<TokenDTO> result = oAuthService.googleLogin(googleDTO);
+        return ResponseEntity.ok(result);
+    }
+
+    /**
+     * 연동여부를 저장하는 API
+     * @param linkYn 연동동의 여부
+     * @return
+     */
+    @PostMapping("/link")
+    public ResponseEntity<?> linkAccount(@RequestParam String email, @RequestParam boolean linkYn) {
+        ResultDTO<SuccessDTO> result = oAuthService.linkAccount(email, linkYn);
         return ResponseEntity.ok(result);
     }
 }
