+++ Back-End/src/main/java/net/scit/backend/oauth/service/OAuthService.java
@@ -1,6 +1,7 @@
 package net.scit.backend.oauth.service;
 
 import net.scit.backend.common.ResultDTO;
+import net.scit.backend.common.SuccessDTO;
 import net.scit.backend.member.dto.TokenDTO;
 import net.scit.backend.oauth.dto.GoogleDTO;
 import org.springframework.stereotype.Service;
@@ -8,4 +9,6 @@ import org.springframework.stereotype.Service;
 @Service
 public interface OAuthService {
     ResultDTO<TokenDTO> googleLogin(GoogleDTO googleDTO);
+
+    ResultDTO<SuccessDTO> linkAccount(String email, boolean linkYn);
 }
